ref(collections): Rename dropFirstWhile to dropWhile
