+++ collections/README.md
@@ -104,26 +104,26 @@ Returns all distinct elements in the given array, preserving order by first
 occurence.
 
 ```ts
-import { dropFirstWhile } from "https://deno.land/std@$STD_VERSION/collections/mod.ts";
-
-const numbers = [3, 2, 5, 2, 5];
-const dropFirstWhileNumbers = dropFirstWhile(numbers, (i) => i !== 2);
-
-console.assert(distinctNumbers === [5, 2, 5]);
-```
-
-### dropFirstWhile
-
-Drop all elements in the given collection until the first element that does not
-match the given predicate.
-
-```ts
 import { distinct } from "https://deno.land/std@$STD_VERSION/collections/mod.ts";
 
 const numbers = [3, 2, 5, 2, 5];
 const distinctNumbers = distinct(numbers);
 
 console.assert(distinctNumbers === [3, 2, 5]);
+```
+
+### dropWhile
+
+Drop all elements in the given collection until the first element that does not
+match the given predicate.
+
+```ts
+import { dropWhile } from "https://deno.land/std@$STD_VERSION/collections/mod.ts";
+
+const numbers = [3, 2, 5, 2, 5];
+const dropWhileNumbers = dropFirstWhile(numbers, (i) => i !== 2);
+
+console.assert(dropWhileNumbers === [5, 2, 5]);
 ```
 
 ### filterEntries
